deployment/requirements.txt
PyYAML==3.12
fire==0.1.3
